Do "the right things"
used the pass `env` parameter that get handles to the right notebook and
other instances. `IPython` is a global that might not be always here and
is "just" convenience for Interactive poking around.

The `load_ext` is necessary if your modules are correctly defined to
avoid right conditions and load your extension at "the right time", and
not just when the JS arrive on the page (which might be random)